+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Web UI version numbers should always match the corresponding version of LBRY App
 
 ### Fixed
   * Fixed sort by date of published content ([#986](https://github.com/lbryio/lbry-app/issues/986))
+  * Fix night mode start time, set to 9PM (#1050)
   *
 
 ### Deprecated